Add Controller

+++ Persistence/Atividade13/src/main/java/br/com/ifba/service/CursoService.java
@@ -12,22 +12,36 @@ import java.util.List;
  *
  * @author ketli
  */
-public class CursoService implements CursoIService {
+public class CursoService implements CursoIService{
     
     private final CursoIDao cursoDao = new CursoDao();
     
    @Override
-   public Curso save(Curso curso){
-       return cursoDao.save(curso);
+   public Curso save(Curso curso) {
+       
+       if(curso == null){
+           throw new RuntimeException("Dados do curso não preenchidos!");
+       }else if(curso.getId() != null){
+           throw new RuntimeException("Curso já existente no banco de dados!");
+       }else{
+           return cursoDao.save(curso);
+       }
    }
    
    @Override
    public Curso update(Curso curso){
-       return cursoDao.update(curso);    
+       if(curso == null){
+           throw new RuntimeException("Dados do curso não preenchidos!");
+       }
+       return cursoDao.update(curso);
+             
    }
    
    @Override
    public void delete(Curso curso){
+       if(curso == null){
+             throw new RuntimeException("Curso já excluido ou não encontrado!");
+       }
        cursoDao.delete(curso);
    }
    
@@ -38,7 +52,10 @@ public class CursoService implements CursoIService {
    
    @Override
    public Curso findById(Long id){
+       Curso curso = cursoDao.findById(id);
+       if(curso == null){
+            throw new RuntimeException("curso nao encontrado");
+       }
        return cursoDao.findById(id);
-   }
-    
+   }  
 }
